pyproject.toml
[project]
name = "test-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.24.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.1",
    "websockets>=12.0",
    "twilio>=9.0.0",
    "pytest>=9.0.1",
]
